+++ cli/router.php
@@ -21,10 +21,10 @@ use Thruway\Transport\RatchetTransportProvider;
 define('AUTHORIZATION_ENABLE', (bool)getenv('AUTHORIZATION_ENABLE'));
 define('AUTH_METHOD', getenv('AUTH_METHOD'));
 define('JWT_SECRET_KEY', (string)getenv('JWT_SECRET_KEY'));
-define('JWT_ALGOS', explode(',', (string)getenv('JWT_ALGOS')) ?: ['HS256']);
+define('JWT_ALGO', (string)getenv('JWT_ALGO') ?: 'HS256');
 define('REALM', (string)getenv('REALM'));
-define('ALLOW_REALM_AUTOCREATE', (bool)getenv('REALM') ?: false);
-define('THRUWAY_DEBUG_ENABLE', (bool)getenv('THRUWAY_DEBUG_ENABLE') ?: false);
+define('ALLOW_REALM_AUTOCREATE', (bool)getenv('REALM'));
+define('THRUWAY_DEBUG_ENABLE', (bool)getenv('THRUWAY_DEBUG_ENABLE'));
 
 define('REDIS_HOST', (string)getenv('REDIS_HOST') ?: 'redis');
 define('REDIS_PORT', (int)getenv('REDIS_PORT') ?: 6379);
@@ -64,7 +64,7 @@ if (AUTHORIZATION_ENABLE) {
 if (AUTH_METHOD === 'jwt') {
     /**@var  $jwtDecoderFactory AbstractJwtDecoderFactory */
     $jwtDecoderFactory = $di->getService(AbstractJwtDecoderFactory::class);
-    $jwtDecoder = $jwtDecoderFactory->createJwtDecoderAdapter(JWT_SECRET_KEY, JWT_ALGOS);
+    $jwtDecoder = $jwtDecoderFactory->createJwtDecoderAdapter(JWT_SECRET_KEY, JWT_ALGO);
     $jwtValidator = $di->getService(JwtValidatorInterface::class);
     $jwtAuthenticationProvider = new JwtAuthenticationProvider([REALM], $jwtDecoder, $jwtValidator);
     $router->addInternalClient($jwtAuthenticationProvider);