+++ src/FreeElephants/Thruway/Jwt/AbstractJwtDecoderFactory.php
@@ -5,5 +5,5 @@ namespace FreeElephants\Thruway\Jwt;
 abstract class AbstractJwtDecoderFactory
 {
 
-    abstract public function createJwtDecoderAdapter(string $key, array $allowedAlgorithms): JwtDecoderAdapterInterface;
+    abstract public function createJwtDecoderAdapter(string $key, string $algorithm): JwtDecoderAdapterInterface;
 }