+++ tests/FreeElephants/Thruway/Jwt/FirebaseJwtAdapterTest.php
@@ -4,7 +4,6 @@
 namespace FreeElephants\Thruway\Jwt;
 
 
-use FreeElephants\Thruway\Jwt\Exception\InvalidArgumentException;
 use FreeElephants\Thruway\Jwt\Exception\OutOfBoundsException;
 use PHPUnit\Framework\TestCase;
 
@@ -13,9 +12,9 @@ class FirebaseJwtAdapterTest extends TestCase
 
     public function testDecode()
     {
-        $decoder = new FirebaseJwtDecoderAdapter('example_key', ['HS256', 'HS384']);
+        $decoder = new FirebaseJwtDecoderAdapter('example_key', 'HS256');
 
-        $signature = 'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJhdXRoaWQiOiJqb2UiLCJhdXRocm9sZXMiOlsic3Vic2NyaWJlciJdfQ.Lxyy1H3gfs1FV5UJLGxfAYvS1TJeiJhVInu5GIlccg4';
+        $signature = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhdXRoaWQiOiJqb2UiLCJhdXRocm9sZXMiOlsic3Vic2NyaWJlciJdfQ.IhmSdzXm8fzSN04XYoNx3CHdYhPJC9ujta_7R6ubQ-w';
         $expected = new \stdClass();
         $expected->authid = 'joe';
         $expected->authroles = [
@@ -25,15 +24,9 @@ class FirebaseJwtAdapterTest extends TestCase
         $this->assertEquals($expected, $decoder->decode($signature));
     }
 
-    public function testUseEmptyAllowedAlgorithmsListInvalidArgumentException()
-    {
-        $this->expectException(InvalidArgumentException::class);
-        new FirebaseJwtDecoderAdapter('example_key', []);
-    }
-
     public function testSetAllowedAlgorithmsOutOfBoundsException()
     {
         $this->expectException(OutOfBoundsException::class);
-        new FirebaseJwtDecoderAdapter('example_key', ['foo bar']);
+        new FirebaseJwtDecoderAdapter('example_key', 'foo bar');
     }
 }