+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased] 
 
+### Changed
+- Update composer dependencies
+
 ## [0.6.0] 2025-02-10
 
 ### Changed