+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased] 
 
 ### Changed
+- Use firebase/jwt:6, refactor related adapter package
+- JWT_ALGOS (comma separated list) replaces with JWT_ALGO (single string value)
 - Update composer dependencies
 
 ## [0.6.0] 2025-02-10